pyproject.toml
[project]
name = "agvv"
version = "0.1.0"
description = "Deterministic project orchestration engine for AI coding agents."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.15",
    "pydantic>=2.0",
    "python-frontmatter>=1.1",
    "rich>=13.0",
]

[project.scripts]
agvv = "agvv.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
